move arrow_io_source to include; remove direct support in the API; update tests
