+++ cpp/include/cudf/io/datasource.hpp
@@ -216,5 +216,50 @@ class datasource {
   };
 };
 
+/**
+ * @brief Implementation class for reading from an Apache Arrow file. The file
+ * could be a memory-mapped file or other implementation supported by Arrow.
+ */
+class arrow_io_source : public datasource {
+  class arrow_io_buffer : public buffer {
+    std::shared_ptr<arrow::Buffer> arrow_buffer;
+
+   public:
+    explicit arrow_io_buffer(std::shared_ptr<arrow::Buffer> arrow_buffer)
+      : arrow_buffer(arrow_buffer)
+    {
+    }
+    size_t size() const override { return arrow_buffer->size(); }
+    const uint8_t* data() const override { return arrow_buffer->data(); }
+  };
+
+ public:
+  explicit arrow_io_source(std::shared_ptr<arrow::io::RandomAccessFile> file) : arrow_file(file) {}
+
+  std::unique_ptr<buffer> host_read(size_t offset, size_t size) override
+  {
+    auto result = arrow_file->ReadAt(offset, size);
+    CUDF_EXPECTS(result.ok(), "Cannot read file data");
+    return std::make_unique<arrow_io_buffer>(result.ValueOrDie());
+  }
+
+  size_t host_read(size_t offset, size_t size, uint8_t* dst) override
+  {
+    auto result = arrow_file->ReadAt(offset, size, dst);
+    CUDF_EXPECTS(result.ok(), "Cannot read file data");
+    return result.ValueOrDie();
+  }
+
+  size_t size() const override
+  {
+    auto result = arrow_file->GetSize();
+    CUDF_EXPECTS(result.ok(), "Cannot get file size");
+    return result.ValueOrDie();
+  }
+
+ private:
+  std::shared_ptr<arrow::io::RandomAccessFile> arrow_file;
+};
+
 }  // namespace io
 }  // namespace cudf
